Former-commit-id: 95b4137478213d91fa3546645b4aeca7f35ced0b

+++ Sources/Internal/Render/RHI/rhi_ShaderSource.h
@@ -111,7 +111,7 @@ public:
 
 private:
     void _Reset();
-    void _AppendLine(const char* line, uint32 lineLen);
+    void _AppendLine(const char* line, size_t lineLen);
 
     struct
     buf_t
